+++ internal/types/file-info.go
@@ -1,0 +1,6 @@
+package types
+
+type FileInfo struct {
+	FullPath   string
+	NeedRemove bool
+}
