+++ internal/utils/utils.go
@@ -1,0 +1,10 @@
+package utils
+
+func ContainedInArray(array []string, value string) bool {
+	for _, n := range array {
+		if value == n {
+			return true
+		}
+	}
+	return false
+}
